[ofono] ril_config_merge_files => config_merge_files
And moved to the core, to make it usable from RIL independent code.
