+++ ofono/.gitignore
@@ -62,6 +62,7 @@ unit/test-sailfish_cell_info_dbus
 unit/test-sailfish_manager
 unit/test-sailfish_sim_info
 unit/test-sailfish_sim_info_dbus
+unit/test-config
 unit/test-watch
 unit/test-sms-filter
 unit/test-voicecall-filter
